refac: rename v2 files to zarrv2 3
